web-ui/requirements.txt
Flask == 3.1.1
requests == 2.32.4
PyYAML == 6.0.2
Flask-RESTful == 0.3.10